Fix for sales

+++ src/com/niki/domain/interactors/catalog/intake/IntakeContract.java
@@ -5,13 +5,13 @@ import com.niki.domain.entities.Provider;
 public class IntakeContract {
     private int id;
     private Provider provider;
-    private int dateTime;
+    private long dateTime;
 
     public IntakeContract() {
         this(0, null, 0);
     }
 
-    public IntakeContract(int id, Provider provider, int dateTime) {
+    public IntakeContract(int id, Provider provider, long dateTime) {
         this.id = id;
         this.provider = provider;
         this.dateTime = dateTime;
@@ -33,11 +33,11 @@ public class IntakeContract {
         this.provider = provider;
     }
 
-    public int getDateTime() {
+    public long getDateTime() {
         return dateTime;
     }
 
-    public void setDateTime(int dateTime) {
+    public void setDateTime(long dateTime) {
         this.dateTime = dateTime;
     }
 }
